Moved filter classes to new file
